[Edge] Ess.Power: distribution prepare for different handlers
- Prepares Ess.Power for different Power-Distribution-Handlers; this will eventually allow us to implement a v2 that is not based on Linear Equation System
- This code should not break anything in existing implementation
- Introduce Simple-Constraint:
  - only one coefficient
  - coefficent value is '1'
  - value is of type int

+++ io.openems.edge.ess.api/src/io/openems/edge/ess/api/ManagedSymmetricEss.java
@@ -705,7 +705,7 @@ public interface ManagedSymmetricEss extends SymmetricEss {
 	 * @throws OpenemsException on error
 	 */
 	public default Constraint createPowerConstraint(String description, SingleOrAllPhase phase, Pwr pwr,
-			Relationship relationship, double value) throws OpenemsException {
+			Relationship relationship, int value) throws OpenemsException {
 		return this.getPower().createSimpleConstraint(description, this, phase, pwr, relationship, value);
 	}
 
@@ -724,7 +724,7 @@ public interface ManagedSymmetricEss extends SymmetricEss {
 	 * @throws OpenemsException on error
 	 */
 	public default Constraint addPowerConstraint(String description, SingleOrAllPhase phase, Pwr pwr,
-			Relationship relationship, double value) throws OpenemsException {
+			Relationship relationship, int value) throws OpenemsException {
 		return this.getPower().addConstraint(this.createPowerConstraint(description, phase, pwr, relationship, value));
 	}
 
@@ -743,7 +743,7 @@ public interface ManagedSymmetricEss extends SymmetricEss {
 	 * @throws OpenemsException on error
 	 */
 	public default Constraint addPowerConstraintAndValidate(String description, SingleOrAllPhase phase, Pwr pwr,
-			Relationship relationship, double value) throws OpenemsException {
+			Relationship relationship, int value) throws OpenemsException {
 		return this.getPower()
 				.addConstraintAndValidate(this.createPowerConstraint(description, phase, pwr, relationship, value));
 	}
